frontend-streamlit/pyproject.toml
[project]
name = "multiagentes-frontend"
version = "0.1.0"
description = "Streamlit Frontend for Multiagents Project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.34.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "httpx>=0.28.0",
    "pytest-asyncio>=0.23.0"
]
quality = [
    "ruff>=0.9.0",
    "bandit>=1.8.0",
    "pip-audit>=2.7.0"
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "S", "B", "UP"]
ignore = ["S603", "S607", "S404", "S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
